Fix incorrectly using None interpolation type
Note: Original fix ported from Tahoma

This PR fixes an problem in Tahoma that also exists in the Official OpenToonz version noted in issue opentoonz/opentoonz#3168

When a new keyframe is created, the keyframe may inadvertently be created as a None interpolation type. There may be multiple ways to do this, but 1 known way is to create a new keyframe before the 1st keyframe in the column and then copy/paste the newly created keyframe after it. This was due to not updating m_prevType in the next keyframe properly.

A None interpolation keyframe is written to the scene file as <n>...<\n> which can lead to a crash while loading the scene or create strange behavior with the keyframes in the column.

This fix should

Prevent the creation of a None interpolation type keyframe

Throw out None interpolation keyframes when loading the scene in case it encounters one.

Co-Authored-By: manongjohn <19245851+manongjohn@users.noreply.github.com>