Fix copy/paste keyframe causing bad data
This PR fix should stop the generation of <n> or <> keyframe data when saving.

The logic that was there assumed and worked correctly for full keyframes, but not partial keyframes. When copy-pasting partial keyframes, it would set m_prevType on channels that were not keyframed. In addition to that, the newly pasted key's interpolation was being incorrectly updated which would lead to the <n>/<> when pasting before a partial keyframe.

Added logic to only update m_prevTypes if the channel was keyframed on the newly pasted key and the next/previous key also had that channel keyframed.

Co-Authored-By: manongjohn <19245851+manongjohn@users.noreply.github.com>