Fix SVG sequence file loading
This PR fixes an issue when trying to load an SVG file sequence.

Modified it so SVG files are treated as Vector Images instead of Vector
Levels, which allows the SVG files to be detected as a single image file
or a series of files (filename.####.svg or filename_####.svg format).

Ported from https://github.com/tahoma2d/tahoma2d/pull/1212

Co-Authored-By: manongjohn <19245851+manongjohn@users.noreply.github.com>